+++ changelog.rst
@@ -10,6 +10,7 @@ ChangeLog
 - set inheritance to Configurable from b3j0f.annotation.PrivateCallInterceptor.
 - remove decorator module.
 - add foreigns attributes in Configurable which allows to add not specified parameters given by conf resources.
+- add autoconf attribute in Configurable.
 
 0.2.1 (2015/10/29)
 ------------------